LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/picture-source-no-img.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"

    <!-- speculative case in document.write -->
    <picture><source srcset="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=22eb3c21-76a6-469c-bab4-21c7176b775f&amp;encodingcheck=&Gbreve;"></picture>


FAIL Speculative parsing, document.write(): picture-source-no-img Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"
